backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mgx-cli"
version = "0.1.0"
description = "MGX AI Command Line Interface"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "MGX Team", email = "info@mgx.ai" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click",
    "pyyaml",
    "requests",
    "pydantic",
    "metagpt",
    "httpx>=0.28.1",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=7.0.0",
    "coverage>=7.10.7",
    "fastapi>=0.125.0",
    "uvicorn[standard]>=0.39.0",
    "freezegun>=1.5.5",
    "pytest-timeout>=2.4.0",
    "pytest-mock>=3.15.1",
    "aiosqlite>=0.22.0",
]

[project.scripts]
mgx = "mgx_cli.main:cli"

[tool.setuptools.packages.find]
include = ["mgx_cli*", "mgx_agent*", "mgx_observability*"]
